Ignore notifies only for cancel observation requests.
Add also isObserveCancel() to the Request.

+++ demo-apps/cf-plugtest-checker/src/main/java/org/eclipse/californium/plugtests/PlugtestChecker.java
@@ -999,7 +999,7 @@ public class PlugtestChecker {
 			} else if (!invert) {
 				System.out.printf("PASS: Observe (%d)\n",
 						// response.getFirstOption(OptionNumberRegistry.OBSERVE).getIntValue());
-						response.getOptions().getObserve().intValue());
+						response.getOptions().getObserve());
 			} else {
 				System.out.println("PASS: No Observe");
 			}
